Move the 8bit masking to Byte

+++ parser/src/main/kotlin/com/linkedin/dex/spec/EncodedValue.kt
@@ -44,7 +44,7 @@ sealed class EncodedValue {
         val VALUE_BOOLEAN: Byte = 0x1f
 
         fun create(byteBuffer: ByteBuffer): EncodedValue {
-            val argAndType = byteBuffer.get().toInt().toUnsigned8BitInt()
+            val argAndType = byteBuffer.get().toUnsigned8BitInt()
 
             // first three bits are the optional valueArg
             val valueArg = (argAndType ushr 5).toByte()
@@ -127,4 +127,4 @@ sealed class EncodedValue {
     }
 }
 
-private fun Int.toUnsigned8BitInt(): Int = (this and 0xFF)
+private fun Byte.toUnsigned8BitInt(): Int = (this.toInt() and 0xFF)
